chore: use android's settings instead of a custom one
